pyproject.toml
[project]
name = "financial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "pandas>=2.3.2",
]
